reorganize clean command
